% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComplexNumFunctions.R
\name{AngleDegrees}
\alias{AngleDegrees}
\title{Get the angle of a complex number (degrees)}
\usage{
AngleDegrees(x)
}
\arguments{
\item{x}{Input is a single complex number or a matrix of complex numbers.}
}
\description{
Get the angle of a complex number (degrees)
}
