% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComplexNumFunctions.R
\name{DotPlotComplex}
\alias{DotPlotComplex}
\title{Plots complex number(s) in Cartesian space}
\usage{
DotPlotComplex(x)
}
\arguments{
\item{x}{Input is a single complex number or a matrix of complex numbers.}
}
\description{
Plots complex number(s) in Cartesian space
}
