% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopoFunctions.R
\name{get_scalpmap}
\alias{get_scalpmap}
\title{Generate scalp map for topoplot.}
\usage{
get_scalpmap(data, ...)
}
\arguments{
\item{data}{Input EEG}

\item{...}{Additional arguments...}

\item{method}{"biharmonic" or "gam" smooth}

\item{grid_res}{Grid resolution}

\item{interp_limit}{Interpolate up to the "head" or add a "skirt"}

\item{quantity}{Quantity to be plotted. Defaults to "amplitude".}

\item{facets}{Any facets you plan to use}

\item{r}{Size of headshape}
}
\description{
Generate scalp map for topoplot.
}
