% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComplexNumFunctions.R
\name{Magnitude}
\alias{Magnitude}
\title{Get the magnitude of a complex number}
\usage{
Magnitude(x)
}
\arguments{
\item{x}{Input is a single complex number or a matrix of complex numbers.}
}
\description{
Get the magnitude of a complex number
}
