% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeTopoPlot.R
\name{MakeTopoPlot}
\alias{MakeTopoPlot}
\title{Makes topoplots.}
\usage{
MakeTopoPlot(
  data,
  chanLocs,
  time_lim,
  limits,
  grid_res,
  palette,
  interp_limit,
  contour,
  highlights,
  groups
)
}
\arguments{
\item{data}{Input EEG}

\item{chanLocs}{Provides native support for BioSemi 64 channel arrays. Other
arrays may be imported as data frames with columns for channel number, channel
label, and Cartesian x and y coordinates.}

\item{time_lim}{Time point(s) to plot, that is, the time range to average across.
If nont supplied, will average across all time points in the input data.}

\item{limits}{Limits of the fill scale - should be given as a character
vector with two values specifying the start and endpoints e.g. limits =
c(-2,-2). Will ignore anything else. Defaults to the range of the data.}

\item{grid_res}{Resolution of the interpolated grid. Higher = smoother but
slower. Defaults to 200 points per edge.}

\item{palette}{Defaults to RdBu if none supplied. Can be any palette from
RColorBrewer or viridis. If an unsupported palette is specified, switches
to Greens.}

\item{interp_limit}{"skirt" or "head". Defaults to "skirt". "skirt"
interpolates just past the farthest electrode and does not respect the
boundary of the head_shape. "head" interpolates up to the radius of the
plotted head, and moves all electrodes inside the head.}

\item{contour}{Plot contour lines on topography (defaults to TRUE)}

\item{highlights}{Electrodes to highlight (in white).}

\item{groups}{Column name for groups to retain.}
}
\description{
Makes topoplots.
}
