% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopoFunctions.R
\name{select_times}
\alias{select_times}
\title{Select EEG time range.}
\usage{
select_times(data, ...)
}
\arguments{
\item{data}{Input EEG}

\item{time_lim}{A character vector of two numbers indicating the time range
to be selected e.g. c(min, max)}
}
\value{
Data frame with only data from within the specified range.
}
\description{
Select EEG time range.
}
