% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BioSemi}
\alias{BioSemi}
\alias{BioSemi64}
\title{BioSemi Channel X & Y Coordinates}
\format{
A `data.frame` object
}
\usage{
BioSemi64
}
\description{
BioSemi Channel X & Y Coordinates
}
\note{
X and Y Cartesian coordinates for 64 channels.

Other included arrays include BioSemi 16, 32, 128, & 256 channel arrays.

Also included is 64 channel coordinates by A B name vs labels.
}
\keyword{datasets}
